pyproject.toml
[project]
name = "agent-best-version"
description = "Minimal browser automation agent for specific tasks"
version = "0.1.0"
requires-python = ">=3.11,<4.0"
dependencies = [
    "aiohttp>=3.12.0",
    "anyio>=4.9.0",
    "bubus>=1.5.6",
    "httpx>=0.28.1",
    "pydantic>=2.11.5",
    "python-dotenv>=1.0.1",
    "typing-extensions>=4.12.2",
    "uuid7>=0.1.0",
    "openai>=2.7.2,<3.0.0",
    "anthropic>=0.72.1,<1.0.0",
    "cdp-use>=1.4.4",
    "pillow>=11.2.1",
    "markdownify>=1.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

